refactor: Move subsystems to a subsystems package
- Subsystems being in the root is bad

BREAKING CHANGE: yes.
