+++ core/src/main/java/com/rowanmcalpin/nextftc/core/command/groups/ParallelRaceGroup.kt
@@ -18,9 +18,7 @@
 
 package com.rowanmcalpin.nextftc.core.command.groups
 
-import com.rowanmcalpin.nextftc.core.Subsystem
 import com.rowanmcalpin.nextftc.core.command.Command
-import com.rowanmcalpin.nextftc.core.command.CommandManager
 
 /**
  * A [CommandGroup] that runs all of its children simultaneously until one of its children is done,
