+++ core/src/main/java/com/rowanmcalpin/nextftc/core/command/utility/InstantCommand.kt
@@ -18,8 +18,6 @@
 
 package com.rowanmcalpin.nextftc.core.command.utility
 
-import com.rowanmcalpin.nextftc.core.Subsystem
-
 /**
  * This is a LambdaCommand that sets isDone to true instantly. As such, there is no update or stop
  * lambda (since the command finishes instantly). All code should be put in the startLambda.
