+++ core/src/main/java/com/rowanmcalpin/nextftc/core/command/utility/LambdaCommand.kt
@@ -18,7 +18,7 @@
 
 package com.rowanmcalpin.nextftc.core.command.utility
 
-import com.rowanmcalpin.nextftc.core.Subsystem
+import com.rowanmcalpin.nextftc.core.subsystems.Subsystem
 import com.rowanmcalpin.nextftc.core.command.Command
 import java.util.function.Consumer
 import java.util.function.Supplier
