+++ ftc/src/main/java/com/rowanmcalpin/nextftc/ftc/driving/DifferentialArcadeDriverControlled.kt
@@ -18,7 +18,6 @@
 
 package com.rowanmcalpin.nextftc.ftc.driving
 
-import com.rowanmcalpin.nextftc.core.Subsystem
 import com.rowanmcalpin.nextftc.core.command.Command
 import com.rowanmcalpin.nextftc.ftc.gamepad.GamepadEx
 import com.rowanmcalpin.nextftc.ftc.gamepad.Joystick
