+++ ftc/src/main/java/com/rowanmcalpin/nextftc/ftc/hardware/controllables/ResetEncoder.kt
@@ -18,7 +18,6 @@
 
 package com.rowanmcalpin.nextftc.ftc.hardware.controllables
 
-import com.rowanmcalpin.nextftc.core.Subsystem
 import com.rowanmcalpin.nextftc.core.command.Command
 
 /**
