+++ ftc/src/main/java/com/rowanmcalpin/nextftc/ftc/hardware/controllables/SetPower.kt
@@ -18,7 +18,7 @@
 
 package com.rowanmcalpin.nextftc.ftc.hardware.controllables
 
-import com.rowanmcalpin.nextftc.core.Subsystem
+import com.rowanmcalpin.nextftc.core.subsystems.Subsystem
 import com.rowanmcalpin.nextftc.core.command.Command
 
 /**
