+++ hardware/src/main/java/com/rowanmcalpin/nextftc/hardware/RunToVelocity.kt
@@ -18,7 +18,7 @@
 
 package com.rowanmcalpin.nextftc.hardware
 
-import com.rowanmcalpin.nextftc.core.Subsystem
+import com.rowanmcalpin.nextftc.core.subsystems.Subsystem
 import dev.nextftc.nextcontrol.ControlSystem
 import dev.nextftc.nextcontrol.KineticState
