+++ pedro/src/main/java/com/rowanmcalpin/nextftc/pedro/DriverControlled.kt
@@ -18,7 +18,7 @@
 
 package com.rowanmcalpin.nextftc.pedro
 
-import com.rowanmcalpin.nextftc.core.Subsystem
+import com.rowanmcalpin.nextftc.core.subsystems.Subsystem
 import com.rowanmcalpin.nextftc.core.command.Command
 import com.rowanmcalpin.nextftc.ftc.gamepad.GamepadEx
 import com.rowanmcalpin.nextftc.ftc.gamepad.Joystick
