+++ pedro/src/main/java/com/rowanmcalpin/nextftc/pedro/FollowPath.kt
@@ -20,7 +20,7 @@ package com.rowanmcalpin.nextftc.pedro
 
 import com.pedropathing.pathgen.Path
 import com.pedropathing.pathgen.PathChain
-import com.rowanmcalpin.nextftc.core.Subsystem
+import com.rowanmcalpin.nextftc.core.subsystems.Subsystem
 import com.rowanmcalpin.nextftc.core.command.Command
 import com.rowanmcalpin.nextftc.ftc.hardware.Drivetrain
